requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bio==1.7.1
biopython==1.85
biothings_client==0.4.1
bleach==6.2.0
blis==1.2.0
cachetools==5.5.1
camelot-py==1.0.0
Cartopy==0.24.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
country_converter==1.3
coverage==7.8.0
cryptography==44.0.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Elixir==0.7.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
et_xmlfile==2.0.0
executing==2.2.0
fastembed==0.5.1
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
google-auth==2.38.0
google-genai==1.2.0
gprofiler-official==1.0.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
langchain==0.3.19
langchain-core==0.3.37
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.10
language_data==1.3.0
llama-cloud==0.1.13
llama-cloud-services==0.6.1
llama-index==0.12.19
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-fastembed==0.3.0
llama-index-embeddings-ollama==0.5.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-groq==0.3.1
llama-index-llms-ollama==0.5.2
llama-index-llms-openai==0.3.20
llama-index-llms-openai-like==0.3.3
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-qdrant==0.4.3
llama-parse==0.6.1
loguru==0.7.3
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.5.5
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
mygene==3.2.2
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.4.7
onnxruntime==1.20.1
openai==1.63.2
opencage==3.1.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.0
pdf2image==1.17.0
pdfminer.six==20240706
pdfplumber==0.11.5
pdftext==0.6.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
pprintpp==0.4.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycountry-convert==0.7.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.2
pymupdf4llm==0.0.17
pyparsing==3.2.3
pypdf==5.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproj==3.7.1
pyreadline3==3.5.4
pyshp==2.3.1
pytesseract==0.3.13
pytest==8.3.5
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
qdrant-client==1.13.2
RapidFuzz==3.12.1
ratelim==0.1.6
referencing==0.36.2
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
setuptools==75.8.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
striprtf==0.0.26
surya-ocr==0.12.0
sympy==1.13.1
table-ocr==0.2.5
tabula-py==2.10.0
tabulate==0.9.0
Tempita==0.6.0
tenacity==9.0.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unpywall==0.2.3
uri-template==1.3.0
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
win32_setctime==1.2.0
wrapt==1.17.2
xlrd==0.7.1
xlwt==0.7.2
yarl==1.18.3
zstandard==0.23.0
